Rename `axes.py` to `axis.py`
